# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CalculEat API Documentation - OpenAPI 3.0",
    "description": "Sample description.",
    "version": "1.0.10"
  },
  "servers": [
    {
      "url": "http://hostname:7854/"
    }
  ],
...
  "paths": {
...
    "/api/1.0/connection": {
      "get": {
        "tags": [
          "connection"
        ],
        "summary": "List all of the connections in DB.",
        "description": "Fetch all of the connections from DB.",
        "operationId": "listConnections",
        "parameters": [
          {
            "name": "is_accepted",
            "in": "query",
            "description": "The id of dietician that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "f"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionResponseBody"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:connections",
              "read:connections"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "connection"
        ],
        "summary": "Add a new connection to the database",
        "description": "Add a new connection to the database. All of the fields are not needed to create connection.",
        "operationId": "createConnection",
        "requestBody": {
          "description": "Create a new connection in the DB.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionPOSTRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:connections",
              "read:connections"
            ]
          }
        ]
      }
    },
    "/api/1.0/connection/{connection_id}": {
      "delete": {
        "tags": [
          "connection"
        ],
        "summary": "Deletes the connection from the database. This can only be done by the logged in user.",
        "description": "Deletes the specified connection according to the ID.",
        "operationId": "deleteConnection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "The connection_id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection successfuly deleted."
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Connection not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:connections",
              "read:connections"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "connection"
        ],
        "summary": "Update connection",
        "description": "This can only be done by the logged in user. **You can use only fields that you want to update.**",
        "operationId": "updateConnection",
        "requestBody": {
          "description": "Update the given connection.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionPATCHRequestBody"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "The connection_id that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionPATCHResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request"
          },
          "404": {
            "description": "Connection not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:connections",
              "read:connections"
            ]
          }
        ]
      }
    },
    "/api/1.0/meeting": {
      "get": {
        "tags": [
          "meeting"
        ],
        "summary": "List all of the meetings in DB.",
        "description": "Fetch all of the meetings from DB.",
        "parameters": [
          {
            "name": "dietician_user_id",
            "in": "query",
            "description": "The id of dietician that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 36
            }
          },
          {
            "name": "client_user_id",
            "in": "query",
            "description": "The id of dietician that needs to be fetched",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 48
            }
          },
          {
            "name": "creation_time",
            "in": "query",
            "description": "The creation datetime of meeting that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2017-07-21T17:32:28.000Z"
            }
          },
          {
            "name": "meet_link",
            "in": "query",
            "description": "The link of meeting that needs to be fetched",
            "required": false,
            "schema": {
              "type": "string",
              "example": "https://meet.calculeat.com/asda-qwe1as-21"
            }
          }
        ],
        "operationId": "listMeetings",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingResponseBody"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "meeting"
        ],
        "summary": "Add a new meeting to the database",
        "description": "Add a new meeting to the database. All of the fields are not needed to create meeting.",
        "operationId": "createMeeting",
        "requestBody": {
          "description": "Create a new meeting in the DB.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      }
    },
    "/api/1.0/meeting/{meeting_id}": {
      "delete": {
        "tags": [
          "meeting"
        ],
        "summary": "Deletes the meeting from the database. This can only be done by the logged in user.",
        "description": "Deletes the specified meeting according to the ID.",
        "operationId": "deleteMeeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "description": "The meeting_id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting successfuly deleted."
          },
          "404": {
            "description": "Meeting not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "meeting"
        ],
        "summary": "Update meeting",
        "description": "This can only be done by the logged in user. **You can use only fields that you want to update.**",
        "operationId": "updateMeeting",
        "requestBody": {
          "description": "Update the meeting.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingRequestBody"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "description": "The meeting_id that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingPATCHResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request"
          },
          "404": {
            "description": "Meeting not found"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "write:meetings",
              "read:meetings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionPATCHResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "dietician_user_id": {
                "type": "integer",
                "format": "int64",
                "example": 21
              },
              "client_user_id": {
                "type": "integer",
                "format": "int64",
                "example": 10
              },
              "is_accepted": {
                "type": "string",
                "example": "t"
              },
              "CreatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "UpdatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "deleted_on": {
                "type": "string",
                "example": "0001-01-01T00:00:00Z"
              }
            }
          },
          "msg": {
            "type": "string",
            "example": "Successful operation, object has been updated."
          }
        }
      },
      "ConnectionPOSTRequestBody": {
        "type": "object",
        "properties": {
          "target_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 21
          }
        }
      },
      "ConnectionPATCHRequestBody": {
        "type": "object",
        "properties": {
          "target_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "is_accepted": {
            "type": "string",
            "format": "char",
            "example": "f"
          }
        }
      },
      "ConnectionResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "dietician_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "client_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "is_accepted": {
            "type": "string",
            "example": "t"
          },
          "CreatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "UpdatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "deleted_on": {
            "type": "string",
            "example": "0001-01-01T00:00:00Z"
          }
        }
      },
      "MeetingPATCHResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "dietician_user_id": {
                "type": "integer",
                "format": "int64",
                "example": 21
              },
              "client_user_id": {
                "type": "integer",
                "format": "int64",
                "example": 10
              },
              "meet_link": {
                "type": "string",
                "example": "https://meet.calculeat.com/asda-qwe1as-21"
              },
              "CreatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "UpdatedAt": {
                "type": "string",
                "example": "2023-09-10T18:25:55.6236225+03:00"
              },
              "deleted_on": {
                "type": "string",
                "example": "0001-01-01T00:00:00Z"
              }
            }
          },
          "msg": {
            "type": "string",
            "example": "Successful operation, object has been updated."
          }
        }
      },
      "MeetingRequestBody": {
        "type": "object",
        "properties": {
          "target_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "meet_link": {
            "type": "string",
            "example": "https://meet.calculeat.com/asda-qwe1as-21"
          }
        }
      },
      "MeetingResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "dietician_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 21
          },
          "client_user_id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "meet_link": {
            "type": "string",
            "example": "https://meet.calculeat.com/asda-qwe1as-21"
          },
          "CreatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "UpdatedAt": {
            "type": "string",
            "example": "2023-09-10T18:25:55.6236225+03:00"
          },
          "deleted_on": {
            "type": "string",
            "example": "0001-01-01T00:00:00Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}